[controller][schema] Move migration schema normalization into StoreSchemaService
