+++ services/venice-controller/src/test/java/com/linkedin/venice/controller/AbstractTestVeniceParentHelixAdmin.java
@@ -102,9 +102,9 @@ public class AbstractTestVeniceParentHelixAdmin {
     SchemaEntry mockEntry = new SchemaEntry(0, TEST_SCHEMA);
     doReturn(mockEntry).when(internalAdmin).getKeySchema(anyString(), anyString());
     // Outside a store-migration context, schema normalization is a passthrough. Mirror that here so
-    // the mocked internal admin doesn't return null and blank out the value schema during createStore/
+    // the mocked schema service doesn't return null and blank out the value schema during createStore/
     // addValueSchema.
-    when(internalAdmin.normalizeSchemaForMigration(anyString(), anyString(), any()))
+    when(storeSchemaService.normalizeSchemaForMigration(anyString(), anyString(), any()))
         .thenAnswer(invocation -> invocation.getArgument(2));
 
     zkClient = mock(ZkClient.class);
